FB Login 1st attempt

+++ BabiesGrow.py
@@ -34,6 +34,101 @@ def showLogin():
     # Render Login Template return "The current session state is %s" % login_session['state']
     return render_template('login.html', STATE=state)
 
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+
+
+    #Exchange client token for long lived server-side token
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())['web']['app_id']
+    app_secret = json.loads(open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (app_id,app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+    #Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.12/me"
+    #strip expire tag from access token
+    token = result.split(",")[0].split(':')[1].replace('"','')
+
+    url = 'https://graph.facebook.com/v2.12/me?access_token=%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    #print "url sent for API access:%s"% url
+    #print "API JSON result: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    login_session['access_token'] = token
+
+    #Get the user's picture
+    url = 'https://graph.facebook.com/v2.12/me/picture?%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # Check if user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+@app.route('/fbdisconnect', methods=['POST'])
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    access_token = login_session['access_token']
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+    del login_session['username']
+    del login_session['facebook_id']
+    del login_session['user_id']
+    del login_session['email']
+    del login_session['picture']
+    return "You have been logged out"
+
+# Disconnect based on provider
+
+@app.route('disconnect')
+def disconnect():
+    if 'provider' in login_session:
+        if login_session['provider'] == 'google':
+            gdisconnect()
+            del login_session['gplus_id']
+            del login_session['access_token']
+        if login_session['provider'] == 'facebook':
+            fbdisconnect()
+            del login_session['facebook_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        flash("You have successfully been logged out.")
+        return redirect(url_for('showRestaurants'))
+    else:
+        flash("You were not logged in")
+        return redirect(url_for('offering'))
 
 # Validate the state token
 @app.route('/gconnect', methods=['POST'])
@@ -90,6 +185,7 @@ def gconnect():
         return response
 
     # Store access token in the session
+    login_session['provider'] = 'google'
     login_session['access_token'] = credentials.access_token
     login_session['gplus_id'] = gplus_id
 
@@ -227,6 +323,8 @@ def newOffering():
 
 @app.route('/offerings/<int:offering_id>/edit/', methods=['GET', 'POST'])
 def editOffering(offering_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     editedOffering = session.query(Offering).filter_by(id=offering_id).one()
     if request.method == 'POST':
         if request.form['title']:
@@ -238,7 +336,11 @@ def editOffering(offering_id):
 
 @app.route('/offerings/<int:offering_id>/delete/', methods=['GET', 'POST'])
 def deleteOffering(offering_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     offeringToDelete = session.query(Offering).filter_by(id=offering_id).one()
+    if offeringToDelete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete this offering.');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(offeringToDelete)
         session.commit()
